Fix building without perl in the build root
After building extensions, Makefile builds documentation by executing
the new perl. Because we version libperl.so soname, we need to
precreate a symlink from the soname to ./libperl.so so that the new
perl can be executed. Otherwise this happens:

LD_LIBRARY_PATH=/builddir/build/BUILD/perl-5.24.0  ./perl -Ilib -f pod/buildtoc -q
./perl: error while loading shared libraries: libperl.so.5.24: cannot open shared object file: No such file or directory

This did not appear on upgrade from previous perl version because perl
build script sets LD_PRELOAD in addition if libperl.so exists in the system.