Move UNIVERSAL to perl-libs
This module name space with methods is implemented in libperl.so. The
module file only supplies a dummy import() method. The module file
exists for a backward compatiblity. Technically we could separate the
module file from perl-libs, but it carry a documentation. The
documentation is similar to one that can be found in perlobj POD, bud
the UNIVERSAL POD is easier to discover.